Add simple tests for synapse-ci images

+++ docker/scripts/test.sh
@@ -29,7 +29,13 @@ docker run --rm -it --entrypoint python vertexproject/synapse:${TAG} -m synapse.
 dstatus00=$?
 if [ $dstatus00 != "0" ]; then exit 1; fi
 
-# FIXME - Add tests for the synapse-ci images
+docker run --rm -it --entrypoint /usr/bin/xvfb-run vertexproject/synapse-ci:${TAG} --help
+dstatus01=$?
+if [ $dstatus01 != "0" ]; then exit 1; fi
+
+docker run --rm -it --entrypoint python vertexproject/synapse:${TAG} -m synapse.servers.cortex --help
+dstatus02=$?
+if [ $dstatus02 != "0" ]; then exit 1; fi
 
 docker run --rm -d --name test-aha -e "SYN_AHA_AHA_NETWORK=synapse.ci" vertexproject/synapse-aha:${TAG}
 docker run --rm -d --name test-axon vertexproject/synapse-axon:${TAG}
